block: queue work on power efficient wq
Block layer uses workqueues for multiple purposes. There is no real dependency
of scheduling these on the cpu which scheduled them.

On a idle system, it is observed that and idle cpu wakes up many times just to
service this work. It would be better if we can schedule it on a cpu which the
scheduler believes to be the most appropriate one.

This patch replaces normal workqueues with power efficient versions.

Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Tejun Heo <tj@kernel.org>
Signed-off-by: Davide Garberi <dade.garberi@gmail.com>
Signed-off-by: Henrique Pereira <hlcpereira@pixelexperience.org>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: Little-W <1405481963@qq.com>